media_server/server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ssi-camera"
version = "0.0.1"
authors = [
  { name="Novak Boškov", email="novak.boskov@nokia-bell-labs.com" },
]
description = "Forward camera stream to the Web browser."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = ["opencv-python",                
                "flask",
                "gunicorn",
                "gevent",
                "requests",
                "pyzmq",
                "redis"]

[project.scripts]
ssic-server = "ssi_camera.main:main"
ssic-publish = "ssi_camera.camera:boot_cameras"
